pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-mddocgen"
version = "0.1.0"
description = "llm-mddocgen - Large Language Model Markdown Document Generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "tomli>=2.0; python_version < '3.11'",
  "pypdf>=4.0",
]

[project.scripts]
llm-mddocgen = "llm_mddocgen.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
include = [
  "llm_mddocgen/**",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
